tests/wpt/tests/html/canvas/element/drawing-rectangles-to-the-canvas/2d.clearRect.nonfinite.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.clearRect.nonfinite</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.clearRect.nonfinite</h1>
<p class="desc">clearRect() with Infinity/NaN is ignored</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("clearRect() with Infinity/NaN is ignored");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50);
  ctx.clearRect(Infinity, 0, 100, 50);
  ctx.clearRect(-Infinity, 0, 100, 50);
  ctx.clearRect(NaN, 0, 100, 50);
  ctx.clearRect(0, Infinity, 100, 50);
  ctx.clearRect(0, -Infinity, 100, 50);
  ctx.clearRect(0, NaN, 100, 50);
  ctx.clearRect(0, 0, Infinity, 50);
  ctx.clearRect(0, 0, -Infinity, 50);
  ctx.clearRect(0, 0, NaN, 50);
  ctx.clearRect(0, 0, 100, Infinity);
  ctx.clearRect(0, 0, 100, -Infinity);
  ctx.clearRect(0, 0, 100, NaN);
  ctx.clearRect(Infinity, Infinity, 100, 50);
  ctx.clearRect(Infinity, Infinity, Infinity, 50);
  ctx.clearRect(Infinity, Infinity, Infinity, Infinity);
  ctx.clearRect(Infinity, Infinity, 100, Infinity);
  ctx.clearRect(Infinity, 0, Infinity, 50);
  ctx.clearRect(Infinity, 0, Infinity, Infinity);
  ctx.clearRect(Infinity, 0, 100, Infinity);
  ctx.clearRect(0, Infinity, Infinity, 50);
  ctx.clearRect(0, Infinity, Infinity, Infinity);
  ctx.clearRect(0, Infinity, 100, Infinity);
  ctx.clearRect(0, 0, Infinity, Infinity);
  _assertPixel(canvas, 50,25, 0,255,0,255);

});
</script>
