LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/drawing-rectangles-to-the-canvas/2d.clearRect.nonfinite-expected.txt

2d.clearRect.nonfinite
clearRect() with Infinity/NaN is ignored
Actual output:
Expected output:


PASS clearRect() with Infinity/NaN is ignored
